+++ Sources/TuistAutomationTesting/Utilities/MockBuildGraphInspector.swift
@@ -43,4 +43,27 @@ public final class MockBuildGraphInspector: BuildGraphInspecting {
             return []
         }
     }
+
+    public var testableTargetStub: ((Scheme, Graph) -> Target?)?
+    public func testableTarget(scheme: Scheme, graph: Graph) -> Target? {
+        if let testableTargetStub = testableTargetStub {
+            return testableTargetStub(scheme, graph)
+        } else {
+            return Target.test()
+        }
+    }
+
+    public var testableSchemesStub: ((Graph) -> [Scheme])?
+    public func testableSchemes(graph: Graph) -> [Scheme] {
+        if let testableSchemesStub = testableSchemesStub {
+            return testableSchemesStub(graph)
+        } else {
+            return []
+        }
+    }
+
+    public var testSchemesStub: ((Graph) -> [Scheme])?
+    public func testSchemes(graph: Graph) -> [Scheme] {
+        testSchemesStub?(graph) ?? []
+    }
 }