+++ Sources/TuistCore/Models/Product.swift
@@ -20,6 +20,7 @@ public enum Product: String, CustomStringConvertible, CaseIterable, Encodable {
     //    case messagesApplication = "messages_application"
     case messagesExtension = "messages_extension"
     case stickerPackExtension = "sticker_pack_extension"
+    case appClip
 
     public var caseValue: String {
         switch self {
@@ -59,6 +60,8 @@ public enum Product: String, CustomStringConvertible, CaseIterable, Encodable {
             return "stickerPackExtension"
         case .commandLineTool:
             return "commandLineTool"
+        case .appClip:
+            return "appClip"
         }
     }
 
@@ -100,13 +103,15 @@ public enum Product: String, CustomStringConvertible, CaseIterable, Encodable {
             return "sticker pack extension"
         case .commandLineTool:
             return "command line tool"
+        case .appClip:
+            return "appClip"
         }
     }
 
     /// Returns true if the target can be ran.
     public var runnable: Bool {
         switch self {
-        case .app, .commandLineTool:
+        case .app, .appClip, .commandLineTool:
             return true
         default:
             return false
@@ -131,6 +136,7 @@ public enum Product: String, CustomStringConvertible, CaseIterable, Encodable {
             base.append(.stickerPackExtension)
             //            base.append(.messagesApplication)
             base.append(.messagesExtension)
+            base.append(.appClip)
         }
 
         if platform == .tvOS {
@@ -206,6 +212,12 @@ public enum Product: String, CustomStringConvertible, CaseIterable, Encodable {
             return .stickerPack
         case .commandLineTool:
             return .commandLineTool
+        case .appClip:
+            return .onDemandInstallCapableApplication
         }
+    }
+
+    public func canHostTests() -> Bool {
+        [.app, .appClip].contains(self)
     }
 }