+++ Sources/TuistKit/ProjectEditor/ProjectEditorMapper.swift
@@ -99,6 +99,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
                               xcodeProjPath: xcodeProjPath,
                               name: "Manifests",
                               organizationName: nil,
+                              developmentRegion: nil,
                               settings: projectSettings,
                               filesGroup: .group(name: "Manifests"),
                               targets: targets,
@@ -127,12 +128,13 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         }
 
         let manifestTargetNodes = manifestsTargets.map { TargetNode(project: project, target: $0, dependencies: dependencies) }
+        let workspace = Workspace(path: project.path, name: "Manifests", projects: [project.path])
 
         let graph = Graph(
             name: "Manifests",
             entryPath: sourceRootPath,
             entryNodes: manifestTargetNodes,
-            workspace: nil,
+            workspace: workspace,
             projects: [project],
             cocoapods: [],
             packages: [],