+++ Sources/TuistKit/Services/FocusService.swift
@@ -14,28 +14,10 @@ protocol FocusServiceProjectGeneratorFactorying {
 
 final class FocusServiceProjectGeneratorFactory: FocusServiceProjectGeneratorFactorying {
     func generator(sources: Set<String>, xcframeworks: Bool, ignoreCache: Bool) -> Generating {
-        let cacheOutputType: CacheOutputType = xcframeworks ? .xcframework : .framework
-        let cacheConfig: CacheConfig = ignoreCache
-            ? .withoutCaching()
-            : .withCaching(cacheOutputType: cacheOutputType)
-        return Generator(graphMapperProvider: GraphMapperProvider(cacheConfig: cacheConfig, sources: sources))
-    }
-}
-
-enum FocusServiceError: FatalError {
-    case cacheWorkspaceNonSupported
-    var description: String {
-        switch self {
-        case .cacheWorkspaceNonSupported:
-            return "Caching is only supported when focusing on a project. Please, run the command in a directory that contains a Project.swift file."
-        }
-    }
-
-    var type: ErrorType {
-        switch self {
-        case .cacheWorkspaceNonSupported:
-            return .abort
-        }
+        let graphMapperProvider = FocusGraphMapperProvider(cache: !ignoreCache,
+                                                           cacheSources: sources,
+                                                           cacheOutputType: xcframeworks ? .xcframework : .framework)
+        return Generator(graphMapperProvider: graphMapperProvider)
     }
 }
 
@@ -55,9 +37,6 @@ final class FocusService {
 
     func run(path: String?, sources: Set<String>, noOpen: Bool, xcframeworks: Bool, ignoreCache: Bool) throws {
         let path = self.path(path)
-        if isWorkspace(path: path) {
-            throw FocusServiceError.cacheWorkspaceNonSupported
-        }
         let generator = projectGeneratorFactory.generator(sources: sources,
                                                           xcframeworks: xcframeworks,
                                                           ignoreCache: ignoreCache)
@@ -75,9 +54,5 @@ final class FocusService {
         } else {
             return FileHandler.shared.currentPath
         }
-    }
-
-    private func isWorkspace(path: AbsolutePath) -> Bool {
-        manifestLoader.manifests(at: path).contains(.workspace)
     }
 }