+++ TapestryConfig.swift
@@ -4,8 +4,9 @@ let config = TapestryConfig(
     release: Release(
         actions:
         [
-            .pre(tool: "git", arguments: ["checkout", "master"]),
+            .pre(tool: "git", arguments: ["checkout", "main"]),
             .pre(tool: "git", arguments: ["pull"]),
+            .pre(tool: "bundle", arguments: ["install"]),
             .pre(.dependenciesCompatibility([.spm(.all)])),
             .pre(tool: "swift", arguments: ["test"]),
             .pre(tool: "bundle", arguments: ["exec", "rake", "features"]),