+++ Tests/TuistCacheTests/TreeShaker/TreeShakerTests.swift
@@ -1,0 +1,5 @@
+import XCTest
+@testable import TuistCache
+@testable import TuistSupportTesting
+
+final class TreeShakerTests: TuistUnitTestCase {}