+++ Tests/TuistCoreTests/Models/ProductTests .swift
@@ -14,6 +14,7 @@ final class ProductTests: XCTestCase {
         XCTAssertEqual(Product.uiTests.xcodeValue, PBXProductType.uiTestBundle)
         XCTAssertEqual(Product.appExtension.xcodeValue, PBXProductType.appExtension)
         XCTAssertEqual(Product.stickerPackExtension.xcodeValue, PBXProductType.stickerPack)
+        XCTAssertEqual(Product.appClip.xcodeValue, PBXProductType.onDemandInstallCapableApplication)
     }
 
     func test_description() {
@@ -25,6 +26,7 @@ final class ProductTests: XCTestCase {
         XCTAssertEqual(Product.uiTests.description, "ui tests")
         XCTAssertEqual(Product.appExtension.description, "app extension")
         XCTAssertEqual(Product.stickerPackExtension.description, "sticker pack extension")
+        XCTAssertEqual(Product.appClip.description, "appClip")
     }
 
     func test_forPlatform_when_ios() {
@@ -40,6 +42,7 @@ final class ProductTests: XCTestCase {
             .messagesExtension,
             .unitTests,
             .uiTests,
+            .appClip,
         ]
         XCTAssertEqual(Set(got), Set(expected))
     }
@@ -74,7 +77,7 @@ final class ProductTests: XCTestCase {
 
     func test_runnable() {
         Product.allCases.forEach { product in
-            if [.app, .commandLineTool].contains(product) {
+            if [.app, .appClip, .commandLineTool].contains(product) {
                 XCTAssertTrue(product.runnable)
             } else {
                 XCTAssertFalse(product.runnable)