+++ fixtures/app_with_tests/Targets/tvOSFramework/Tests/tvOSFrameworkTests.swift
@@ -1,0 +1,12 @@
+import Foundation
+import XCTest
+
+@testable import tvOSFramework
+
+final class tvOSFrameworkTests: XCTestCase {
+    func testHello() {
+        let sut = tvOSFramework()
+        
+        XCTAssertEqual("tvOSFramework.hello()", sut.hello())
+    }
+}