+++ fixtures/ios_app_with_carthage_frameworks/Core/CoreFile.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public class CoreFile {
-    public init() {}
-
-    public func hello() -> String {
-        return "CoreFile.hello()"
-    }
-}